test(effect-log): the stable-hash tests are a sibling file
The test-layout check (xtask) refuses an inline test module; the
stable_hash tests move to log/stable_hash_tests.rs.
